// frontend/src/api/Capability.ts
export enum Capability {
  BasicControl = 'BasicControlCapability',
  CarpetModeControl = 'CarpetModeControlCapability',
  CombinedVirtualRestrictions = 'CombinedVirtualRestrictionsCapability',
  ConsumableMonitoring = 'ConsumableMonitoringCapability',
  Debug = 'DebugCapability',
  DoNotDisturb = 'DoNotDisturbCapability',
  FanSpeedControl = 'FanSpeedControlCapability',
  GoToLocation = 'GoToLocationCapability',
  LEDControl = 'LEDControlCapability',
  Locate = 'LocateCapability',
  ManualControl = 'ManualControlCapability',
  MapReset = 'MapResetCapability',
  MapSegmentEdit = 'MapSegmentEditCapability',
  MapSegmentRename = 'MapSegmentRenameCapability',
  MapSegmentation = 'MapSegmentationCapability',
  MapSnapshot = 'MapSnapshotCapability',
  PersistentMapControl = 'PersistentMapControlCapability',
  SensorCalibration = 'SensorCalibrationCapability',
  SpeakerTest = 'SpeakerTestCapability',
  SpeakerVolumeControl = 'SpeakerVolumeControlCapability',
  VoicePackManagement = 'VoicePackManagementCapability',
  WaterUsageControl = 'WaterUsageControlCapability',
  WifiConfiguration = 'WifiConfigurationCapability',
  ZoneCleaning = 'ZoneCleaningCapability',
}